requirements.txt
djongo==1.3.4
djangorestframework==3.12.4
requests==2.25.1